requirements-update.txt
finnhub-python
requests